Restore game_mode.h
