+++ src/common/utils/excluded_apps.h
@@ -1,17 +1,6 @@
 #pragma once
-#include <shellapi.h>
 #include <vector>
 #include <string>
-
-inline bool detect_game_mode()
-{
-    QUERY_USER_NOTIFICATION_STATE notification_state;
-    if (SHQueryUserNotificationState(&notification_state) != S_OK)
-    {
-        return false;
-    }
-    return (notification_state == QUNS_RUNNING_D3D_FULL_SCREEN);
-}
 
 // Checks if a process path is included in a list of strings.
 bool find_app_name_in_path(const std::wstring& where, const std::vector<std::wstring>& what)
